+++ src/editor/dataset/constant/Checkbox.ts
@@ -6,5 +6,5 @@ export const defaultCheckboxOption: Readonly<Required<ICheckboxOption>> = {
   gap: 5,
   lineWidth: 1,
   fillStyle: '#5175f4',
-  fontStyle: '#ffffff'
+  strokeStyle: '#ffffff'
 }