+++ src/editor/dataset/constant/Separator.ts
@@ -1,0 +1,6 @@
+import { ISeparatorOption } from '../../interface/Separator'
+
+export const defaultSeparatorOption: Readonly<Required<ISeparatorOption>> = {
+  lineWidth: 1,
+  strokeStyle: '#000000'
+}