+++ src/editor/dataset/enum/Row.ts
@@ -2,5 +2,6 @@ export enum RowFlex {
   LEFT = 'left',
   CENTER = 'center',
   RIGHT = 'right',
-  ALIGNMENT = 'alignment'
+  ALIGNMENT = 'alignment',
+  JUSTIFY = 'justify'
 }