+++ src/editor/interface/Control.ts
@@ -2,6 +2,7 @@ import { ControlType, ControlIndentation } from '../dataset/enum/Control'
 import { EditorZone } from '../dataset/enum/Editor'
 import { ICheckbox } from './Checkbox'
 import { IElement } from './Element'
+import { IRadio } from './Radio'
 import { IRange } from './Range'
 
 export interface IValueSet {
@@ -20,6 +21,12 @@ export interface IControlCheckbox {
   max?: number
   valueSets: IValueSet[]
   checkbox?: ICheckbox
+}
+
+export interface IControlRadio {
+  code: string | null
+  valueSets: IValueSet[]
+  radio?: IRadio
 }
 
 export interface IControlHighlightRule {
@@ -65,6 +72,7 @@ export type IControl = IControlBasic &
   IControlRule &
   Partial<IControlSelect> &
   Partial<IControlCheckbox> &
+  Partial<IControlRadio> &
   Partial<IControlStyle>
 
 export interface IControlOption {