+++ src/editor/interface/Position.ts
@@ -9,6 +9,7 @@ export interface ICurrentPosition {
   x?: number
   y?: number
   isCheckbox?: boolean
+  isRadio?: boolean
   isControl?: boolean
   isImage?: boolean
   isTable?: boolean
@@ -41,6 +42,7 @@ export type IGetFloatPositionByXYPayload = IGetPositionByXYPayload & {
 export interface IPositionContext {
   isTable: boolean
   isCheckbox?: boolean
+  isRadio?: boolean
   isControl?: boolean
   index?: number
   trIndex?: number