+++ src/editor/interface/Separator.ts
@@ -1,0 +1,4 @@
+export interface ISeparatorOption {
+  strokeStyle?: string
+  lineWidth?: number
+}